feat: default to using override-flagomatic flag
This makes the override-flagomatic flag enabled in the ebuild by
default, without changing the default behavior. This was done by adding
LTO_ENABLE_FLAGOMATIC=yes to make.conf.lto. This allows the user to now
use the flag-o-matic.conf file to disable flagomatic instead of using it
to enable it.

Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>